Remove legacy generate.go file
